59295112 ruby2 0 sketchup

+++ openstudiocore/src/utilities/core/Containers.hpp
@@ -53,7 +53,7 @@ UTILITIES_API std::vector<std::string> eraseEmptyElements(const std::vector<std:
 template<typename T, typename U> 
 std::vector<T> castVector(const std::vector<U>& objects) {
   std::vector<T> result;
-  for (unsigned i = 0, n = objects.size(); i < n; ++i) {
+  for (size_t i = 0, n = objects.size(); i < n; ++i) {
     T castObject = objects[i].template cast<T>();
     result.push_back( castObject );
   }
